pyproject.toml
[tool.poetry]
name = "coinlist-python"
version = "0.1.0"
description = "This is an unofficial python wrapper for the Coinlist exchange"
authors = ["dmitrii"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9.5"
python-dotenv = "^0.17.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
